+++ include/low_level/potentiometermotor.h
@@ -8,34 +8,34 @@ class PotentiometerMotor {
 public:
   PotentiometerMotor(ShiftRegister *m_shiftRegister, const bool m_backwards,
                      const uint8_t m_pwmPin, const uint8_t m_adcChannel,
-                     const uint8_t m_directionBit, const float m_Kp,
-                     const float m_Ki, const float m_Kd, const char m_ID);
+                     const uint8_t m_directionBit, const double m_Kp,
+                     const double m_Ki, const double m_Kd, const char m_ID);
 
   void init(void);
 
   void update(void);
 
-  void setAngle(float angle);
+  void setAngle(double angle);
 
-  float getAngle(void);
+  double getAngle(void);
 
-  inline void setPID(float m_Kp, float m_Ki, float m_Kd) {
+  inline void setPID(double m_Kp, double m_Ki, double m_Kd) {
     Kp = m_Kp;
     Ki = m_Kp;
     Kd = m_Kp;
   }
 
-  inline float getError() {return error;}
-  inline float getSetPoint() {return targetAngle;}
-  inline float getP() {return P;}
-  inline float getI() {return I;}
-  inline float getD() {return D;}
+  inline double getError() {return error;}
+  inline double getSetPoint() {return targetAngle;}
+  inline double getP() {return P;}
+  inline double getI() {return I;}
+  inline double getD() {return D;}
 
-  inline void setMaxCE(float max_ce) {maxCE = max_ce;}
+  inline void setMaxCE(double max_ce) {maxCE = max_ce;}
 
 private:
-  float calculatePID(void);
-  void setPWM(float dutyCycle, uint8_t m_direction);
+  double calculatePID(void);
+  void setPWM(double dutyCycle, uint8_t m_direction);
 
   const char ID;
 
@@ -45,14 +45,14 @@ private:
 
   ShiftRegister *const shiftRegister;
   int backwards = 1;
-  float currentAngle, targetAngle;
-  float error = 0, previousError, cumulativeError = 0;
+  double currentAngle, targetAngle;
+  double error = 0, previousError, cumulativeError = 0;
   uint64_t timeLastUpdated;
   uint32_t deltaT;
   uint8_t direction = 0;
-  float Kp, Ki, Kd;
-  float maxCE = 0;
-  float P, I, D = 0.0;
+  double Kp, Ki, Kd;
+  double maxCE = 0;
+  double P, I, D = 0.0;
 };
 
 #endif
